python/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.12.0
arxiv==1.4.8
scholarly==0.5.1
pydantic==2.5.2
langchain==0.1.0
langchain-openai==0.0.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1 